backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.0
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
apscheduler==3.10.4
pydantic==2.5.3
numpy==1.26.3
